+++ storm-core/src/clj/backtype/storm/daemon/supervisor.clj
@@ -69,7 +69,7 @@
   (map-val :master-code-dir assignments-snapshot))
 
 (defn- read-downloaded-storm-ids [conf]
-  (map #(java.net.URLDecoder/decode %) (read-dir-contents (supervisor-stormdist-root conf)))
+  (map #(url-decode %) (read-dir-contents (supervisor-stormdist-root conf)))
   )
 
 (defn read-worker-heartbeat [conf id]
@@ -438,32 +438,64 @@
       (FileUtils/moveDirectory (File. tmproot) (File. stormroot))
       ))
 
+(defn jlp [stormroot conf]
+  (let [resource-root (str stormroot File/separator RESOURCES-SUBDIR)
+        os (clojure.string/replace (System/getProperty "os.name") #"\s+" "_")
+        arch (System/getProperty "os.arch")
+        arch-resource-root (str resource-root File/separator os "-" arch)]
+    (str arch-resource-root File/pathSeparator resource-root File/pathSeparator (conf JAVA-LIBRARY-PATH)))) 
+
+(defn- substitute-worker-childopts [value port]
+  (let [sub-fn (fn [s] (.replaceAll s "%ID%" (str port)))]
+    (if (list? value)
+      (map sub-fn value)
+      (-> value sub-fn (.split " ")))))
+
+(defn java-cmd []
+  (let [java-home (.get (System/getenv) "JAVA_HOME")]
+    (if (nil? java-home)
+      "java"
+      (str java-home file-path-separator "bin" file-path-separator "java")
+      )))
+
 
 (defmethod launch-worker
     :distributed [supervisor storm-id port worker-id]
     (let [conf (:conf supervisor)
           storm-home (System/getProperty "storm.home")
           stormroot (supervisor-stormdist-root conf storm-id)
+          jlp (jlp stormroot conf)
           stormjar (supervisor-stormjar-path stormroot)
           storm-conf (read-supervisor-storm-conf conf storm-id)
           classpath (add-to-classpath (current-classpath) [stormjar])
-          childopts (.replaceAll (str (conf WORKER-CHILDOPTS) " " (storm-conf TOPOLOGY-WORKER-CHILDOPTS))
-                                 "%ID%"
-                                 (str port))
+          worker-childopts (when-let [s (conf WORKER-CHILDOPTS)]
+                             (substitute-worker-childopts s port))
+          topo-worker-childopts (when-let [s (storm-conf TOPOLOGY-WORKER-CHILDOPTS)]
+                                  (substitute-worker-childopts s port))
           logfilename (str "worker-" port ".log")
-          command (str "java -server " childopts
-                       " -Djava.library.path=" (conf JAVA-LIBRARY-PATH)
-                       " -Dlogfile.name=" logfilename
-                       " -Dstorm.home=" storm-home
-                       " -Dlogback.configurationFile=" storm-home "/logback/cluster.xml"
-                       " -Dstorm.id=" storm-id
-                       " -Dworker.id=" worker-id
-                       " -Dworker.port=" port
-                       " -cp " classpath " backtype.storm.daemon.worker "
-                       (java.net.URLEncoder/encode storm-id) " " (:assignment-id supervisor)
-                       " " port " " worker-id)]
-      (log-message "Launching worker with command: " command)
-      (launch-process command :environment {"LD_LIBRARY_PATH" (conf JAVA-LIBRARY-PATH)})
+          command (concat
+                    [(java-cmd) "-server"]
+                    worker-childopts
+                    topo-worker-childopts
+                    [(str "-Djava.library.path=" jlp)
+                     (str "-Dlogfile.name=" logfilename)
+                     (str "-Dstorm.home=" storm-home)
+                     (str "-Dlogback.configurationFile=" storm-home "/logback/cluster.xml")
+                     (str "-Dstorm.id=" storm-id)
+                     (str "-Dworker.id=" worker-id)
+                     (str "-Dworker.port=" port)
+                     "-cp" classpath
+                     "backtype.storm.daemon.worker"
+                     storm-id
+                     (:assignment-id supervisor)
+                     port
+                     worker-id])
+          command (->> command (map str) (filter (complement empty?)))
+          shell-cmd (->> command
+                         (map #(str \' (clojure.string/escape % {\' "\\'"}) \'))
+                         (clojure.string/join " "))]
+      (log-message "Launching worker with command: " shell-cmd)
+      (launch-process command :environment {"LD_LIBRARY_PATH" jlp})
       ))
 
 ;; local implementation