+++ storm-core/src/jvm/backtype/storm/utils/ShellProcess.java
@@ -17,133 +17,128 @@
  */
 package backtype.storm.utils;
 
+import backtype.storm.Config;
+import backtype.storm.multilang.ISerializer;
+import backtype.storm.multilang.BoltMsg;
+import backtype.storm.multilang.NoOutputException;
+import backtype.storm.multilang.ShellMsg;
+import backtype.storm.multilang.SpoutMsg;
 import backtype.storm.task.TopologyContext;
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.DataOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.util.Map;
+import java.io.InputStream;
+import java.io.Serializable;
 import java.util.List;
+import java.util.Map;
 
 import org.apache.commons.io.IOUtils;
-import org.json.simple.JSONObject;
-import org.json.simple.JSONValue;
 import org.apache.log4j.Logger;
 
-public class ShellProcess {
+public class ShellProcess implements Serializable {
     public static Logger LOG = Logger.getLogger(ShellProcess.class);
-    private DataOutputStream processIn;
-    private BufferedReader processOut;
-    private InputStream processErrorStream;
-    private Process _subprocess;
-    private String[] command;
+    public static Logger ShellLogger;
+    private Process      _subprocess;
+    private InputStream  processErrorStream;
+    private String[]     command;
+    public ISerializer   serializer;
 
     public ShellProcess(String[] command) {
         this.command = command;
     }
 
-    public Number launch(Map conf, TopologyContext context) throws IOException {
+    public Number launch(Map conf, TopologyContext context) {
         ProcessBuilder builder = new ProcessBuilder(command);
         builder.directory(new File(context.getCodeDir()));
-        _subprocess = builder.start();
 
-        processIn = new DataOutputStream(_subprocess.getOutputStream());
-        processOut = new BufferedReader(new InputStreamReader(_subprocess.getInputStream()));
-        processErrorStream = _subprocess.getErrorStream();
+        ShellLogger = Logger.getLogger(context.getThisComponentId());
 
-        JSONObject setupInfo = new JSONObject();
-        setupInfo.put("pidDir", context.getPIDDir());
-        setupInfo.put("conf", conf);
-        setupInfo.put("context", context);
-        writeMessage(setupInfo);
+        this.serializer = getSerializer(conf);
 
-        return (Number)readMessage().get("pid");
+        Number pid;
+        try {
+            _subprocess = builder.start();
+            processErrorStream = _subprocess.getErrorStream();
+            serializer.initialize(_subprocess.getOutputStream(), _subprocess.getInputStream());
+            pid = serializer.connect(conf, context);
+        } catch (IOException e) {
+            throw new RuntimeException(
+                    "Error when launching multilang subprocess\n"
+                            + getErrorsString(), e);
+        } catch (NoOutputException e) {
+            throw new RuntimeException(e + getErrorsString() + "\n");
+        }
+        return pid;
+    }
+
+    private ISerializer getSerializer(Map conf) {
+        //get factory class name
+        String serializer_className = (String)conf.get(Config.TOPOLOGY_MULTILANG_SERIALIZER);
+        LOG.info("Storm multilang serializer: " + serializer_className);
+
+        ISerializer serializer = null;
+        try {
+            //create a factory class
+            Class klass = Class.forName(serializer_className);
+            //obtain a serializer object
+            Object obj = klass.newInstance();
+            serializer = (ISerializer)obj;
+        } catch(Exception e) {
+            throw new RuntimeException("Failed to construct multilang serializer from serializer " + serializer_className, e);
+        }
+        return serializer;
     }
 
     public void destroy() {
         _subprocess.destroy();
     }
 
-    public void writeMessage(Object msg) throws IOException {
-        writeString(JSONValue.toJSONString(msg));
+    public ShellMsg readShellMsg() throws IOException {
+        try {
+            return serializer.readShellMsg();
+        } catch (NoOutputException e) {
+            throw new RuntimeException(e + getErrorsString() + "\n");
+        }
     }
 
-    private void writeString(String str) throws IOException {
-        byte[] strBytes = str.getBytes("UTF-8");
-        processIn.write(strBytes, 0, strBytes.length);
-        processIn.writeBytes("\nend\n");
-        processIn.flush();
+    public void writeBoltMsg(BoltMsg msg) throws IOException {
+        serializer.writeBoltMsg(msg);
+        // Log any info sent on the error stream
+        logErrorStream();
     }
 
-    public JSONObject readMessage() throws IOException {
-        String string = readString();
-        JSONObject msg = (JSONObject)JSONValue.parse(string);
-        if (msg != null) {
-            return msg;
-        } else {
-            throw new IOException("unable to parse: " + string);
+    public void writeSpoutMsg(SpoutMsg msg) throws IOException {
+        serializer.writeSpoutMsg(msg);
+        // Log any info sent on the error stream
+        logErrorStream();
+    }
+
+    public void writeTaskIds(List<Integer> taskIds) throws IOException {
+        serializer.writeTaskIds(taskIds);
+        // Log any info sent on the error stream
+        logErrorStream();
+    }
+
+    public void logErrorStream() {
+        try {
+            while (processErrorStream.available() > 0) {
+                int bufferSize = processErrorStream.available();
+                byte[] errorReadingBuffer = new byte[bufferSize];
+                processErrorStream.read(errorReadingBuffer, 0, bufferSize);
+                ShellLogger.info(new String(errorReadingBuffer));
+            }
+        } catch (Exception e) {
         }
     }
 
     public String getErrorsString() {
-        if(processErrorStream!=null) {
+        if (processErrorStream != null) {
             try {
                 return IOUtils.toString(processErrorStream);
-            } catch(IOException e) {
+            } catch (IOException e) {
                 return "(Unable to capture error stream)";
             }
         } else {
             return "";
         }
-    }
-
-    public void drainErrorStream()
-    {
-        try {
-            while (processErrorStream.available() > 0)
-            {
-                int bufferSize = processErrorStream.available();
-                byte[] errorReadingBuffer =  new byte[bufferSize];
-
-                processErrorStream.read(errorReadingBuffer, 0, bufferSize);
-
-                LOG.info("Got error from shell process: " + new String(errorReadingBuffer));
-            }
-        } catch(Exception e) {
-        }
-    }
-
-    private String readString() throws IOException {
-        StringBuilder line = new StringBuilder();
-
-        //synchronized (processOut) {
-            while (true) {
-                String subline = processOut.readLine();
-                if(subline==null) {
-                    StringBuilder errorMessage = new StringBuilder();
-                    errorMessage.append("Pipe to subprocess seems to be broken!");
-                    if (line.length() == 0) {
-                        errorMessage.append(" No output read.\n");
-                    }
-                    else {
-                        errorMessage.append(" Currently read output: " + line.toString() + "\n");
-                    }
-                    errorMessage.append("Shell Process Exception:\n");
-                    errorMessage.append(getErrorsString() + "\n");
-                    throw new RuntimeException(errorMessage.toString());
-                }
-                if(subline.equals("end")) {
-                    break;
-                }
-                if(line.length()!=0) {
-                    line.append("\n");
-                }
-                line.append(subline);
-            }
-            //}
-
-        return line.toString();
     }
 }