+++ storm-core/test/clj/backtype/storm/messaging/netty_unit_test.clj
@@ -37,7 +37,8 @@
         server (.bind context nil port)
         client (.connect context nil "localhost" port)
         _ (.send client task (.getBytes req_msg))
-        resp (.recv server 0)]
+        iter (.recv server 0 0)
+        resp (.next iter)]
     (is (= task (.task resp)))
     (is (= req_msg (String. (.message resp))))
     (.close client)
@@ -58,7 +59,8 @@
         server (.bind context nil port)
         client (.connect context nil "localhost" port)
         _ (.send client task (.getBytes req_msg))
-        resp (.recv server 0)]
+        iter (.recv server 0 0)
+        resp (.next iter)]
     (is (= task (.task resp)))
     (is (= req_msg (String. (.message resp))))
     (.close client)
@@ -77,15 +79,23 @@
                     }
         context (TransportFactory/makeContext storm-conf)
         client (.connect context nil "localhost" port)
+        
+        server (Thread.
+                (fn []
+                  (Thread/sleep 1000)
+                  (let [server (.bind context nil port)
+                        iter (.recv server 0 0)
+                        resp (.next iter)]
+                    (is (= task (.task resp)))
+                    (is (= req_msg (String. (.message resp))))
+                    (.close server) 
+                  )))
+        _ (.start server)
         _ (.send client task (.getBytes req_msg))
-        _ (Thread/sleep 1000)
-        server (.bind context nil port)
-        resp (.recv server 0)]
-    (is (= task (.task resp)))
-    (is (= req_msg (String. (.message resp))))
+        ]
     (.close client)
-    (.close server)
-    (.term context)))    
+    (.join server)
+    (.term context)))
 
 (deftest test-batch
   (let [storm-conf {STORM-MESSAGING-TRANSPORT "backtype.storm.messaging.netty.Context"
@@ -102,11 +112,21 @@
     (doseq [num  (range 1 100000)]
       (let [req_msg (str num)]
         (.send client task (.getBytes req_msg))))
-    (doseq [num  (range 1 100000)]
+    
+    (let [resp (ArrayList.)
+          received (atom 0)]
+      (while (< @received (- 100000 1))
+        (let [iter (.recv server 0 0)]
+          (while (.hasNext iter)
+            (let [msg (.next iter)]
+              (.add resp msg)
+              (swap! received inc)
+              ))))
+      (doseq [num  (range 1 100000)]
       (let [req_msg (str num)
-            resp (.recv server 0)
-            resp_msg (String. (.message resp))]
-        (is (= req_msg resp_msg))))
+            resp_msg (String. (.message (.get resp (- num 1))))]
+        (is (= req_msg resp_msg)))))
+   
     (.close client)
     (.close server)
     (.term context)))