+++ CHANGELOG.md
@@ -1,3 +1,67 @@
+## 0.9.2-incubating (unreleased)
+ * STORM-335: add drpc test for removing timed out requests from queue
+ * STORM-69: Storm UI Visualizations for Topologies
+ * STORM-297: Performance scaling with CPU
+ * STORM-244: DRPC timeout can return null instead of throwing an exception
+ * STORM-63: remove timeout drpc request from its function's request queue
+ * STORM-313: Remove log-level-page from logviewer
+ * STORM-205: Add REST API To Storm UI
+ * STORM-326: tasks send duplicate metrics
+ * STORM-331: Update the Kafka dependency of storm-kafka to 0.8.1.1
+ * STORM-308: Add support for config_value to {supervisor,nimbus,ui,drpc,logviewer} childopts
+ * STORM-309: storm-starter Readme: windows documentation update
+ * STORM-318: update storm-kafka to use apache curator-2.4.0
+ * STORM-303: storm-kafka reliability improvements
+ * STORM-233: Removed inline heartbeat to nimbus to avoid workers being killed when under heavy ZK load
+ * STORM-267: fix package name of LoggingMetricsConsumer in storm.yaml.example
+ * STORM-265: upgrade to clojure 1.5.1
+ * STORM-232: ship JNI dependencies with the topology jar
+ * STORM-295: Add storm configuration to define JAVA_HOME
+ * STORM-138: Pluggable serialization for multilang
+ * STORM-264: Removes references to the deprecated topology.optimize
+ * STORM-245: implement Stream.localOrShuffle() for trident
+ * STORM-317: Add SECURITY.md to release binaries
+ * STORM-310: Change Twitter authentication
+ * STORM-305: Create developer documentation
+ * STORM-280: storm unit tests are failing on windows
+ * STORM-298: Logback file does not include full path for metrics appender fileNamePattern
+ * STORM-316: added validation to registermetrics to have timebucketSizeInSecs >= 1
+ * STORM-315: Added progress bar when submitting topology
+ * STORM-214: Windows: storm.cmd does not properly handle multiple -c arguments
+ * STORM-306: Add security documentation
+ * STORM-302: Fix Indentation for pom.xml in storm-dist
+ * STORM-235: Registering a null metric should blow up early
+ * STORM-113: making thrift usage thread safe for local cluster
+ * STORM-223: use safe parsing for reading YAML
+ * STORM-238: LICENSE and NOTICE files are duplicated in storm-core jar
+ * STORM-276: Add support for logviewer in storm.cmd.
+ * STORM-286: Use URLEncoder#encode with the encoding specified.
+ * STORM-296: Storm kafka unit tests are failing on windows
+ * STORM-291: upgrade http-client to 4.3.3
+ * STORM-252: Upgrade curator to latest version
+ * STORM-294: Commas not escaped in command line
+ * STORM-287: Fix the positioning of documentation strings in clojure code
+ * STORM-290: Fix a log binding conflict caused by curator dependencies
+ * STORM-289: Fix Trident DRPC memory leak
+ * STORM-173: Treat command line "-c" option number config values as such
+ * STORM-194: Support list of strings in *.worker.childopts, handle spaces
+ * STORM-288: Fixes version spelling in pom.xml
+ * STORM-208: Add storm-kafka as an external module
+ * STORM-285: Fix storm-core shade plugin config
+ * STORM-12: reduce thread usage of netty transport
+ * STORM-281: fix and issue with config parsing that could lead to leaking file descriptors
+ * STORM-196: When JVM_OPTS are set, storm jar fails to detect storm.jar from environment
+ * STORM-260: Fix a potential race condition with simulated time in Storm's unit tests
+ * STORM-258: Update commons-io version to 2.4
+ * STORM-270: don't package .clj files in release jars.
+ * STORM-273: Error while running storm topologies on Windows using "storm jar"
+ * STROM-247: Replace links to github resources in storm script
+ * STORM-263: Update Kryo version to 2.21+
+ * STORM-187: Fix Netty error "java.lang.IllegalArgumentException: timeout value is negative"
+ * STORM-186: fix float secs to millis long convertion
+ * STORM-70: Upgrade to ZK-3.4.5 and curator-1.3.3
+ * STORM-146: Unit test regression when storm is compiled with 3.4.5 zookeeper
+
 ## 0.9.1-incubating
 * Fix to prevent Nimbus from hanging if random data is sent to nimbus thrift port
 * Improved support for running on Windows platforms